% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2013_Accident_Documentation.R
\docType{data}
\name{accident_2013_test.csv}
\alias{accident_2013_test.csv}
\title{2013 Fatal Car Accidents Statistics (FCARS)}
\format{
A data frame with 30202 observations and 50 variables:
\describe{
  \item{STATE}{state identifier..from 1-52}
  \item{ST_CASE}{assigned case number}
  \item{PERSONS}{number of persons fatally injured in the accident}
  \item{LATITUDE}{number of persons fatally injured in the accident}
  \item{LONGITUD}{number of persons fatally injured in the accident}
}
}
\source{
\url{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}
}
\usage{
accident_2013_test.csv
}
\description{
A dataset containing the fatal accident statistics on roads in the United States for 2013
}
\keyword{datasets}
