% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_test.R
\name{fnAdd}
\alias{fnAdd}
\title{Add together two numbers}
\usage{
fnAdd(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers
}
\examples{
fnAdd(1, 1)

}
